+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Changes
 * Added option to hide dashboard time picker
 * Added Notification for Alert
 * Added alertRuleTags field to the graph panel
+* Added support for thresholds to graph panel
 * Added support for Elasticsearch alert condition
 * Added support for using gridPos for dashboard panels